Refactor: Consolidate wallet connection modals and share ConnectEco component
This commit refactors the wallet connection modal flow to reduce code duplication and improve maintainability.

Key changes include:
- Deleted `PriorityWalletConnectModal` and consolidated its functionality into `WalletSelectorModal`.
- Introduced a `showPriorityView` prop in `WalletSelectorModal` to conditionally render the priority connection UI (listing wallets relevant to the source asset) or the standard wallet list.
- Extracted the duplicated UI logic for displaying ecosystem connection rows (Cosmos, EVM, SVM) from the old modals into a new shared component, `ConnectEco`.
- Updated `ConnectedWalletModal` and `WalletSelectorModal` (in priority view) to use the shared `ConnectEco` component.
- Updated `SwapPage` to call the appropriate modal (`WalletSelectorModal` with `showPriorityView: true` or `ConnectedWalletModal`) based on whether a source asset is selected.
- Removed redundant comments

This consolidation eliminates the need for a separate priority modal and ensures the ecosystem row UI is maintained in a single location.
